requirements-minimal.txt
boto3==1.40.55
python-dotenv==1.1.1
locust==2.42.0
sentence-transformers==5.1.1
sagemaker==2.253.1
requests-aws4auth==1.3.1
opensearch-py==3.0.0
"optimum[onnxruntime]"